voice-interview-system/whisper_worker/requirements.txt
openai-whisper
redis
